@@ -5,8 +5,7 @@
   "author": "Matters <hi@matters.news>",
   "main": "build/index.js",
   "engines": {
-    "node": ">=12.16",
-    "npm": ">=6.14"
+    "node": ">=16"
   },
   "license": "Apache-2.0",
   "scripts": {
@@ -77,84 +76,81 @@
     "graphql-tools": "^4.0.8",
     "graphql-upload": "^11.0.0",
     "helmet": "^3.23.3",
-    "ipfs-http-client": "^41.0.0",
+    "ipfs-http-client": "^50.1.2",
     "js-base64": "^3.4.5",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.19.5",
     "lodash": "^4.17.21",
-    "mime-types": "^2.1.27",
+    "mime-types": "^2.1.31",
     "module-alias": "^2.2.2",
-    "nanoid": "^3.1.11",
+    "nanoid": "^3.1.23",
     "number-precision": "^1.5.0",
     "oauth2-server": "^3.1.1",
-    "objection": "^2.2.2",
+    "objection": "^2.2.15",
     "passport": "^0.4.1",
-    "passport-oauth2": "^1.5.0",
+    "passport-oauth2": "^1.6.0",
     "pg": "^8.3.0",
     "psl": "^1.8.0",
     "qs": "^6.9.4",
     "request-ip": "^2.1.3",
-    "stripe": "^8.144.0",
-    "typescript": "^3.8.3",
-    "utf-8-validate": "^5.0.2",
-    "uuid": "^8.3.0",
-    "validator": "^13.1.1",
+    "stripe": "^8.161.0",
+    "typescript": "^4.3.5",
+    "uuid": "^8.3.2",
+    "validator": "^13.6.0",
     "winston": "^3.3.3",
     "winston-transport": "^4.4.0",
     "xss": "^1.0.9"
   },
   "devDependencies": {
-    "@types/bcrypt": "^3.0.0",
-    "@types/bull": "^3.14.0",
-    "@types/cheerio": "^0.22.21",
-    "@types/cookie": "^0.4.0",
+    "@types/bcrypt": "^5.0.0",
+    "@types/bull": "^3.15.2",
+    "@types/cookie": "^0.4.1",
     "@types/dotenv": "^8.2.0",
-    "@types/express": "4.17.1",
-    "@types/express-serve-static-core": "^4.17.1",
     "@types/graphql": "^14.5.0",
     "@types/graphql-depth-limit": "^1.1.2",
     "@types/graphql-relay": "^0.6.0",
+    "@types/graphql-upload": "^8.0.6",
     "@types/helmet": "0.0.47",
     "@types/ioredis": "^4.17.3",
-    "@types/jest": "^26.0.5",
+    "@types/jest": "^26.0.24",
     "@types/js-base64": "^3.0.0",
-    "@types/jsonwebtoken": "^8.5.0",
+    "@types/jsonwebtoken": "^8.5.4",
     "@types/knex": "^0.16.1",
-    "@types/lodash": "^4.14.158",
+    "@types/lodash": "^4.14.171",
     "@types/mime-types": "^2.1.0",
     "@types/nanoid": "^2.1.0",
-    "@types/node-fetch": "^2.5.7",
-    "@types/oauth2-server": "^3.0.12",
-    "@types/passport": "^1.0.4",
-    "@types/passport-oauth2": "^1.4.9",
+    "@types/node-fetch": "^2.5.11",
+    "@types/oauth2-server": "^3.0.13",
+    "@types/passport": "^1.0.7",
+    "@types/passport-oauth2": "^1.4.11",
     "@types/psl": "^1.1.0",
     "@types/qs": "^6.9.4",
     "@types/request-ip": "0.0.35",
-    "@types/uuid": "^8.0.1",
-    "@types/validator": "^13.1.0",
+    "@types/uuid": "^8.3.1",
+    "@types/validator": "^13.6.3",
     "@types/winston": "^2.4.4",
-    "@types/ws": "^7.2.6",
+    "@types/ws": "^7.4.6",
     "apollo": "^2.30.1",
     "apollo-server-testing": "^2.9.12",
-    "chokidar": "^3.4.2",
+    "chokidar": "^3.5.2",
     "chokidar-cli": "^2.1.0",
-    "codecov": "^3.7.2",
-    "cz-conventional-changelog": "^3.2.0",
-    "husky": "^4.2.5",
-    "jest": "^25.5.4",
+    "codecov": "^3.8.2",
+    "cz-conventional-changelog": "^3.3.0",
+    "husky": "^7.0.1",
+    "jest": "^27.0.6",
     "knex-migrate": "^1.7.4",
-    "nodemon": "^2.0.4",
+    "nodemon": "^2.0.12",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.0.5",
-    "rimraf": "^3.0.1",
-    "ts-jest": "^25.5.1",
-    "ts-node": "^8.10.2",
+    "prettier": "^2.3.2",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^27.0.3",
+    "ts-node": "^10.1.0",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0"
   },
   "optionalDependencies": {
-    "bufferutil": "^4.0.1",
-    "utf-8-validate": "^5.0.2"
+    "bufferutil": "^4.0.3",
+    "utf-8-validate": "^5.0.5"
   },
   "_moduleAliases": {
     "common": "./build/common",
