@@ -8,66 +8,13 @@ exports.up = async (knex) => {
     .first()
 
   const mattyTagIds = [
-    6898,
-    6899,
-    7114,
-    10167,
-    10446,
-    11631,
-    11633,
-    11708,
-    11838,
-    12778,
-    13004,
-    13006,
-    13145,
-    13699,
-    13796,
-    14631,
-    15017,
-    16421,
-    16679,
-    17273,
-    17503,
-    19034,
-    19296,
-    19497,
-    19756,
-    19862,
-    20833,
-    21371,
+    6898, 6899, 7114, 10167, 10446, 11631, 11633, 11708, 11838, 12778, 13004,
+    13006, 13145, 13699, 13796, 14631, 15017, 16421, 16679, 17273, 17503, 19034,
+    19296, 19497, 19756, 19862, 20833, 21371,
 
-    1211,
-    3937,
-    6025,
-    7727,
-    9549,
-    9913,
-    10249,
-    10431,
-    10537,
-    10716,
-    10989,
-    11168,
-    11354,
-    11492,
-    11528,
-    11650,
-    11705,
-    12691,
-    13482,
-    13704,
-    13801,
-    14105,
-    14428,
-    14547,
-    14551,
-    15137,
-    15493,
-    15608,
-    16209,
-    18832,
-    21292,
+    1211, 3937, 6025, 7727, 9549, 9913, 10249, 10431, 10537, 10716, 10989,
+    11168, 11354, 11492, 11528, 11650, 11705, 12691, 13482, 13704, 13801, 14105,
+    14428, 14547, 14551, 15137, 15493, 15608, 16209, 18832, 21292,
   ]
 
   for (const id of mattyTagIds) {
