@@ -183,7 +183,8 @@ export const APPRECIATION_TYPES = {
 }
 
 export const SEARCH_KEY_TRUNCATE_LENGTH = 100
-export const SEARCH_ARTICLE_URL_REGEX = /^(https:\/\/([a-z0-9-]+.)?matters.news\/)@([a-zA-Z0-9_-]+)\/(.+?)-([0-9a-zA-Z]{49,59})$/gi
+export const SEARCH_ARTICLE_URL_REGEX =
+  /^(https:\/\/([a-z0-9-]+.)?matters.news\/)@([a-zA-Z0-9_-]+)\/(.+?)-([0-9a-zA-Z]{49,59})$/gi
 
 export const OAUTH_PROVIDER = {
   facebbook: 'facebook',
