@@ -94,10 +94,8 @@ test('correctNestedBrTag', async () => {
       toBe: '<p><s><u><br class="smart"><br class="smart"></u></s></p>',
     },
     {
-      data:
-        '<p><em><s><u><br class="smart"><br class="smart" /></br></u></s></em></p>',
-      toBe:
-        '<p><em><s><u><br class="smart"><br class="smart"></u></s></em></p>',
+      data: '<p><em><s><u><br class="smart"><br class="smart" /></br></u></s></em></p>',
+      toBe: '<p><em><s><u><br class="smart"><br class="smart"></u></s></em></p>',
     },
     {
       data: '<p><br class="smart"><br class="smart"/></br></p>',
@@ -108,27 +106,20 @@ test('correctNestedBrTag', async () => {
       toBe: '<p><u><br class="smart"><br class="smart"/></br></u></p>',
     },
     {
-      data:
-        '<p><br class="smart"><br class="smart"><br class="smart" /></br></br></p>',
+      data: '<p><br class="smart"><br class="smart"><br class="smart" /></br></br></p>',
       toBe: '<p><br class="smart"><br class="smart"><br class="smart"></p>',
     },
     {
-      data:
-        '<p><u><br class="smart"><br class="smart"><br class="smart" /></br></br></u></p>',
-      toBe:
-        '<p><u><br class="smart"><br class="smart"><br class="smart"></u></p>',
+      data: '<p><u><br class="smart"><br class="smart"><br class="smart" /></br></br></u></p>',
+      toBe: '<p><u><br class="smart"><br class="smart"><br class="smart"></u></p>',
     },
     {
-      data:
-        '<blockquote><br class="smart"><br class="smart"><br class="smart" /></br></br></blockquote>',
-      toBe:
-        '<blockquote><br class="smart"><br class="smart"><br class="smart"></blockquote>',
+      data: '<blockquote><br class="smart"><br class="smart"><br class="smart" /></br></br></blockquote>',
+      toBe: '<blockquote><br class="smart"><br class="smart"><br class="smart"></blockquote>',
     },
     {
-      data:
-        '<blockquote><u><br class="smart"><br class="smart"><br class="smart" /></br></br></u></blockquote>',
-      toBe:
-        '<blockquote><u><br class="smart"><br class="smart"><br class="smart"></u></blockquote>',
+      data: '<blockquote><u><br class="smart"><br class="smart"><br class="smart" /></br></br></u></blockquote>',
+      toBe: '<blockquote><u><br class="smart"><br class="smart"><br class="smart"></u></blockquote>',
     },
     {
       data: '<p><br class="not-smart"><br class="not-smart" /></br></p>',
@@ -147,27 +138,20 @@ test('correctNestedBrTag', async () => {
       toBe: '<p><u><br class="smart">A<br class="smart"></u></p>',
     },
     {
-      data:
-        '<p><br class="smart">A<br class="smart">B<br class="smart" /></br></br></p>',
+      data: '<p><br class="smart">A<br class="smart">B<br class="smart" /></br></br></p>',
       toBe: '<p><br class="smart">A<br class="smart">B<br class="smart"></p>',
     },
     {
-      data:
-        '<p><u><br class="smart">A<br class="smart">B<br class="smart" /></br></br></u></p>',
-      toBe:
-        '<p><u><br class="smart">A<br class="smart">B<br class="smart"></u></p>',
+      data: '<p><u><br class="smart">A<br class="smart">B<br class="smart" /></br></br></u></p>',
+      toBe: '<p><u><br class="smart">A<br class="smart">B<br class="smart"></u></p>',
     },
     {
-      data:
-        '<p><br class="smart">A is something... <br class="smart">B<br class="smart" /></br></br></p>',
-      toBe:
-        '<p><br class="smart">A is something... <br class="smart">B<br class="smart"></p>',
+      data: '<p><br class="smart">A is something... <br class="smart">B<br class="smart" /></br></br></p>',
+      toBe: '<p><br class="smart">A is something... <br class="smart">B<br class="smart"></p>',
     },
     {
-      data:
-        '<p><u><br class="smart">A is something... <br class="smart">B<br class="smart" /></br></br></u></p>',
-      toBe:
-        '<p><u><br class="smart">A is something... <br class="smart">B<br class="smart"></u></p>',
+      data: '<p><u><br class="smart">A is something... <br class="smart">B<br class="smart" /></br></br></u></p>',
+      toBe: '<p><u><br class="smart">A is something... <br class="smart">B<br class="smart"></u></p>',
     },
     {
       data:
@@ -186,28 +170,20 @@ test('correctNestedBrTag', async () => {
         '<p><u><br class="smart">A is something... <br class="smart"><em>B</em>C<br class="smart"></u></p>',
     },
     {
-      data:
-        '<blockquote><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart" /></br></br></blockquote>',
-      toBe:
-        '<blockquote><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart"></blockquote>',
+      data: '<blockquote><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart" /></br></br></blockquote>',
+      toBe: '<blockquote><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart"></blockquote>',
     },
     {
-      data:
-        '<blockquote><em><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart" /></br></br></em></blockquote>',
-      toBe:
-        '<blockquote><em><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart"></em></blockquote>',
+      data: '<blockquote><em><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart" /></br></br></em></blockquote>',
+      toBe: '<blockquote><em><br class="smart">A is something.<br class="smart"><em>B</em>C<br class="smart"></em></blockquote>',
     },
     {
-      data:
-        '<p><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart" /></br></p>',
-      toBe:
-        '<p><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart"></p>',
+      data: '<p><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart" /></br></p>',
+      toBe: '<p><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart"></p>',
     },
     {
-      data:
-        '<p><s><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart" /></br></s></p>',
-      toBe:
-        '<p><s><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart"></s></p>',
+      data: '<p><s><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart" /></br></s></p>',
+      toBe: '<p><s><br class="smart">A is <a href="matters.news" target="_blank">here</a>.<br class="smart"></s></p>',
     },
     {
       data:
@@ -246,10 +222,8 @@ test('correctNestedBrTag', async () => {
         '<p><s>A<br class="smart">B<br class="smart" />C<br class="smart" /></br></s><br class="smart">D</p>',
     },
     {
-      data:
-        '<p><a href="" target="_blank"><em><s><u>A<br class="smart"><br class="smart">o</u></s></em></a> B<br class="smart"></p>',
-      toBe:
-        '<p><a href="" target="_blank"><em><s><u>A<br class="smart"><br class="smart">o</u></s></em></a> B<br class="smart"></p>',
+      data: '<p><a href="" target="_blank"><em><s><u>A<br class="smart"><br class="smart">o</u></s></em></a> B<br class="smart"></p>',
+      toBe: '<p><a href="" target="_blank"><em><s><u>A<br class="smart"><br class="smart">o</u></s></em></a> B<br class="smart"></p>',
     },
   ]
 
