@@ -4030,9 +4030,7 @@ export interface GQLArticleTypeResolver<TParent = any> {
   relatedArticles?: ArticleToRelatedArticlesResolver<TParent>
   relatedDonationArticles?: ArticleToRelatedDonationArticlesResolver<TParent>
   appreciationsReceived?: ArticleToAppreciationsReceivedResolver<TParent>
-  appreciationsReceivedTotal?: ArticleToAppreciationsReceivedTotalResolver<
-    TParent
-  >
+  appreciationsReceivedTotal?: ArticleToAppreciationsReceivedTotalResolver<TParent>
   subscribers?: ArticleToSubscribersResolver<TParent>
   appreciateLimit?: ArticleToAppreciateLimitResolver<TParent>
   appreciateLeft?: ArticleToAppreciateLeftResolver<TParent>
@@ -5088,22 +5086,12 @@ export interface GQLNotificationSettingTypeResolver<TParent = any> {
   mention?: NotificationSettingToMentionResolver<TParent>
   userNewFollower?: NotificationSettingToUserNewFollowerResolver<TParent>
   articleNewComment?: NotificationSettingToArticleNewCommentResolver<TParent>
-  articleNewAppreciation?: NotificationSettingToArticleNewAppreciationResolver<
-    TParent
-  >
-  articleNewSubscription?: NotificationSettingToArticleNewSubscriptionResolver<
-    TParent
-  >
-  articleSubscribedNewComment?: NotificationSettingToArticleSubscribedNewCommentResolver<
-    TParent
-  >
-  articleCommentPinned?: NotificationSettingToArticleCommentPinnedResolver<
-    TParent
-  >
+  articleNewAppreciation?: NotificationSettingToArticleNewAppreciationResolver<TParent>
+  articleNewSubscription?: NotificationSettingToArticleNewSubscriptionResolver<TParent>
+  articleSubscribedNewComment?: NotificationSettingToArticleSubscribedNewCommentResolver<TParent>
+  articleCommentPinned?: NotificationSettingToArticleCommentPinnedResolver<TParent>
   circleNewFollower?: NotificationSettingToCircleNewFollowerResolver<TParent>
-  circleNewDiscussion?: NotificationSettingToCircleNewDiscussionResolver<
-    TParent
-  >
+  circleNewDiscussion?: NotificationSettingToCircleNewDiscussionResolver<TParent>
 }
 
 export interface NotificationSettingToEnableResolver<
@@ -5242,9 +5230,7 @@ export interface GQLRecommendationTypeResolver<TParent = any> {
   following?: RecommendationToFollowingResolver<TParent>
   followeeArticles?: RecommendationToFolloweeArticlesResolver<TParent>
   followeeComments?: RecommendationToFolloweeCommentsResolver<TParent>
-  followeeDonatedArticles?: RecommendationToFolloweeDonatedArticlesResolver<
-    TParent
-  >
+  followeeDonatedArticles?: RecommendationToFolloweeDonatedArticlesResolver<TParent>
   followingTagsArticles?: RecommendationToFollowingTagsArticlesResolver<TParent>
   newest?: RecommendationToNewestResolver<TParent>
   hottest?: RecommendationToHottestResolver<TParent>
@@ -7772,9 +7758,7 @@ export interface GQLUserActivityTypeResolver<TParent = any> {
   appreciationsSent?: UserActivityToAppreciationsSentResolver<TParent>
   appreciationsSentTotal?: UserActivityToAppreciationsSentTotalResolver<TParent>
   appreciationsReceived?: UserActivityToAppreciationsReceivedResolver<TParent>
-  appreciationsReceivedTotal?: UserActivityToAppreciationsReceivedTotalResolver<
-    TParent
-  >
+  appreciationsReceivedTotal?: UserActivityToAppreciationsReceivedTotalResolver<TParent>
 }
 
 export interface UserActivityToHistoryArgs {
@@ -9505,9 +9489,7 @@ export interface GQLMutationTypeResolver<TParent = any> {
   userLogout?: MutationToUserLogoutResolver<TParent>
   generateLikerId?: MutationToGenerateLikerIdResolver<TParent>
   updateUserInfo?: MutationToUpdateUserInfoResolver<TParent>
-  updateNotificationSetting?: MutationToUpdateNotificationSettingResolver<
-    TParent
-  >
+  updateNotificationSetting?: MutationToUpdateNotificationSettingResolver<TParent>
   toggleFollowUser?: MutationToToggleFollowUserResolver<TParent>
   toggleBlockUser?: MutationToToggleBlockUserResolver<TParent>
   toggleSubscribePush?: MutationToToggleSubscribePushResolver<TParent>
