@@ -24,14 +24,13 @@ export const inRecommendIcymi: ArticleOSSToInRecommendIcymiResolver = async (
   { dataSources: { articleService } }
 ) => !!(await articleService.findRecommendIcymi(articleId))
 
-export const inRecommendHottest: ArticleOSSToInRecommendHottestResolver = async (
-  { articleId },
-  _,
-  { dataSources: { articleService } }
-) => {
-  const recommendSetting = await articleService.findRecommendSetting(articleId)
-  return recommendSetting.inHottest
-}
+export const inRecommendHottest: ArticleOSSToInRecommendHottestResolver =
+  async ({ articleId }, _, { dataSources: { articleService } }) => {
+    const recommendSetting = await articleService.findRecommendSetting(
+      articleId
+    )
+    return recommendSetting.inHottest
+  }
 
 export const inRecommendNewest: ArticleOSSToInRecommendNewestResolver = async (
   { articleId },
