@@ -1,5 +1,5 @@
 import bodyParser from 'body-parser'
-import { Router } from 'express'
+import { RequestHandler, Router } from 'express'
 import _ from 'lodash'
 import Stripe from 'stripe'
 
@@ -27,7 +27,7 @@ const stripeRouter = Router()
  *
  * @see {@url https://stripe.com/docs/webhooks}
  */
-stripeRouter.use(bodyParser.raw({ type: 'application/json' }))
+stripeRouter.use(bodyParser.raw({ type: 'application/json' }) as RequestHandler)
 
 stripeRouter.post('/', async (req, res) => {
   const paymentService = new PaymentService()
